ToPut/kivy-pipette-master/requirements.txt
kivy
https://github.com/kivymd/KivyMD/archive/master.zip
pillow